+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 -   Improved `os.installPackage()` to support upgrading and downgrading packages.
     -   When downgrading a package, the `downgrade` property on the `options` parameter object should be set to `true`. This is a safety feature to help prevent accidentally downgrading a package version.
 -   Added the ability to "login to a studio" via custom domains.
-    -   Users will recieve a new account when logging into CasualOS served through a custom domain that is scoped to the studio that owns the custom domain.
+    -   Users will receive a new account when logging into CasualOS served through a custom domain that is scoped to the studio that owns the custom domain.
     -   This provides an additional privacy boundary for users, helping ensure that data isn't automatically shared with custom domains simply by logging in.
 
 ### :bug: Bug Fixes